requirements.txt
streamlit
numpy
opencv-python
scikit-learn
pillow
matplotlib
torch
gtts
ipython
scikit-neuralnetwork
streamlit-drawable-canvas
skorch